+++ natural_language_processing/g2p_en/g2p_en_model.cpp
@@ -84,9 +84,9 @@ void forward(AILIANetwork *ailia, std::vector<AILIATensor*> &inputs, std::vector
 		setErrorDetail("input blob cnt and input tensor size must be same", "");
 	}
 
-	for (int i = 0; i < inputs.size(); i++){
+	for (size_t i = 0; i < inputs.size(); i++){
 		unsigned int input_blob_idx = 0;
-		status = ailiaGetBlobIndexByInputIndex(ailia, &input_blob_idx, i);
+		status = ailiaGetBlobIndexByInputIndex(ailia, &input_blob_idx, (unsigned int)i);
 		if (status != AILIA_STATUS_SUCCESS) {
 			setErrorDetail("ailiaGetBlobIndexByInputIndex", ailiaGetErrorDetail(ailia));
 		}
@@ -100,7 +100,7 @@ void forward(AILIANetwork *ailia, std::vector<AILIATensor*> &inputs, std::vector
 			setErrorDetail("ailiaSetInputBlobShape",ailiaGetErrorDetail(ailia));
 		}
 
-		status = ailiaSetInputBlobData(ailia, &(inputs[i]->data)[0], inputs[i]->data.size() * sizeof(float), input_blob_idx);
+		status = ailiaSetInputBlobData(ailia, &(inputs[i]->data)[0], (unsigned int)(inputs[i]->data.size() * sizeof(float)), input_blob_idx);
 		if (status != AILIA_STATUS_SUCCESS) {
 			setErrorDetail("ailiaSetInputBlobData",ailiaGetErrorDetail(ailia));
 		}
@@ -117,9 +117,9 @@ void forward(AILIANetwork *ailia, std::vector<AILIATensor*> &inputs, std::vector
 		setErrorDetail("ailiaGetOutputBlobCount",ailiaGetErrorDetail(ailia));
 	}
 
-	for (int i = 0; i < output_blob_cnt; i++){
+	for (size_t i = 0; i < output_blob_cnt; i++){
 		unsigned int output_blob_idx = 0;
-		status = ailiaGetBlobIndexByOutputIndex(ailia, &output_blob_idx, i);
+		status = ailiaGetBlobIndexByOutputIndex(ailia, &output_blob_idx, (unsigned int)i);
 		if (status != AILIA_STATUS_SUCCESS) {
 			setErrorDetail("ailiaGetBlobIndexByInputIndex",ailiaGetErrorDetail(ailia));
 		}
@@ -141,12 +141,12 @@ void forward(AILIANetwork *ailia, std::vector<AILIATensor*> &inputs, std::vector
 		
 		AILIATensor &ref_tensor = outputs[i];
 		int new_shape = output_blob_shape.x*output_blob_shape.y*output_blob_shape.z*output_blob_shape.w;
-		if (new_shape != ref_tensor.data.size()){
+		if ((size_t)new_shape != ref_tensor.data.size()){
 			ref_tensor.data.resize(new_shape);
 		}
 		ref_tensor.shape = output_blob_shape;
 
-		status = ailiaGetBlobData(ailia, &ref_tensor.data[0], ref_tensor.data.size() * sizeof(float), output_blob_idx);
+		status = ailiaGetBlobData(ailia, &ref_tensor.data[0], (unsigned int)(ref_tensor.data.size() * sizeof(float)), output_blob_idx);
 		if (status != AILIA_STATUS_SUCCESS) {
 			setErrorDetail("ailiaGetBlobData",ailiaGetErrorDetail(ailia));
 		}
@@ -277,7 +277,7 @@ std::vector<int> tokenize(const std::string& word) {
 	
 	std::map<std::string, int> g2idx;
 	for(size_t i = 0; i < graphemes.size(); ++i) {
-		g2idx[graphemes[i]] = i;
+		g2idx[graphemes[i]] = (int)i;
 	}
 
 	std::vector<int> x;
@@ -308,7 +308,7 @@ std::vector<std::string> get_phonemes() {
 std::unordered_map<int, std::string> get_idx2p(const std::vector<std::string>& phonemes) {
 	std::unordered_map<int, std::string> idx2p;
 	for (size_t i = 0; i < phonemes.size(); ++i) {
-		idx2p[i] = phonemes[i];
+		idx2p[(int)i] = phonemes[i];
 	}
 	return idx2p;
 }
@@ -329,7 +329,7 @@ std::vector<std::string> G2PEnModel::predict(const std::string &word){
 	std::vector<int> x = tokenize(word);
 	if (debug){
 		PRINT_OUT("tokens : ");
-		for (int i = 0; i < x.size(); i++){
+		for (size_t i = 0; i < x.size(); i++){
 			PRINT_OUT("%d ", x[i]);
 		}
 		PRINT_OUT("\n");
@@ -345,9 +345,9 @@ std::vector<std::string> G2PEnModel::predict(const std::string &word){
 	h_tensor.shape.w = 1;
 	h_tensor.shape.dim = 2;
 
-	for (int i = 0; i < x.size(); i++){
+	for (size_t i = 0; i < x.size(); i++){
 		std::vector<float> x_data(1);
-		x_data[0] = x[i];
+		x_data[0] = (float)x[i];
 
 		AILIATensor x_tensor;
 		x_tensor.data = x_data;
@@ -372,11 +372,11 @@ std::vector<std::string> G2PEnModel::predict(const std::string &word){
 
 	for (int i = 0; i < 20; i++){
 		std::vector<float> pred_data(1);
-		pred_data[0] = pred;
+		pred_data[0] = (float)pred;
 
 		AILIATensor pred_tensor;
 		pred_tensor.data = pred_data;
-		pred_tensor.shape.x = pred_data.size();
+		pred_tensor.shape.x = (unsigned int)pred_data.size();
 		pred_tensor.shape.y = 1;
 		pred_tensor.shape.z = 1;
 		pred_tensor.shape.w = 1;
@@ -393,10 +393,10 @@ std::vector<std::string> G2PEnModel::predict(const std::string &word){
 		h_tensor = decoder_outputs[1];
 
 		float max_logits = -1;
-		for (int i = 0; i < logits_tensor.shape.x; i++){
+		for (size_t i = 0; i < logits_tensor.shape.x; i++){
 			if (max_logits < logits_tensor.data[i]){
 				max_logits = logits_tensor.data[i];
-				pred = i;
+				pred = (int)i;
 			}
 		}
 
@@ -409,7 +409,7 @@ std::vector<std::string> G2PEnModel::predict(const std::string &word){
 
 	if (debug){
 		PRINT_OUT("output\n");
-		for (int i = 0; i < preds.size(); i++){
+		for (size_t i = 0; i < preds.size(); i++){
 			PRINT_OUT("%d ", preds[i]);
 		}
 		PRINT_OUT("\n");
@@ -515,7 +515,7 @@ std::vector<std::string> G2PEnModel::compute(std::string text)
 			pron = predict(word);
 		}
 
-		for (int i = 0; i < pron.size(); i++) {
+		for (size_t i = 0; i < pron.size(); i++) {
 			prons.push_back(pron[i]);
 		}
 		prons.push_back(" ");
